requirements.txt
MySQL-python>=1.2.3
funcsigs==1.0.2
mock==2.0.0
pbr==1.10.0
pkg-resources==0.0.0
python-dateutil==2.6.0
requests==2.12.1
requests-mock==1.1.0
six==1.10.0
